migrate-worker/requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
attrs==26.1.0
croniter==6.2.2
greenlet==3.5.0
packaging==26.2
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
pydantic==2.13.3
pydantic_core==2.46.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
SQLAlchemy==2.0.49
sqlmodel==0.0.38
typing-inspection==0.4.2
typing_extensions==4.15.0
